Add Policy

+++ app/Http/Controllers/MahasiswaController.php
@@ -31,6 +31,7 @@ class MahasiswaController extends Controller
      */
     public function store(Request $request)
     {
+        $this->authorize('create', Mahasiswa::class);
         $validasi = $request->validate([
             "npm"=> "required|unique:mahasiswas",
             "nama"=> "required",
@@ -60,6 +61,7 @@ class MahasiswaController extends Controller
      */
     public function edit(Mahasiswa $mahasiswa)
     {
+        $this->authorize('update', $mahasiswa);
         $prodi = Prodi::all();
         return view("mahasiswa.edit")->with("mahasiswa",$mahasiswa)->with("prodi", $prodi);
     }
@@ -69,6 +71,7 @@ class MahasiswaController extends Controller
      */
     public function update(Request $request, $id)
     {
+        $this->authorize('update', Mahasiswa::class);
         $validasi = $request->validate([
             "npm" => "required",
             "nama" => "required",
@@ -92,6 +95,7 @@ class MahasiswaController extends Controller
      */
     public function destroy(Mahasiswa $mahasiswa)
     {
+        $this->authorize('delete', $mahasiswa);
         $mahasiswa->delete();
         return redirect()->route("mahasiswa.index")->with("success","Data berhasil dihapus");
     }
